@@ -3,6 +3,7 @@
   "version": "3.1.0",
   "description": "print the filename and line number of the caller (use only in development environment)",
   "main": "index.js",
+  "types": "index.d.ts",
   "repository": "https://github.com/pinojs/pino-caller.git",
   "author": "phra <greensoncio@gmail.com>",
   "license": "MIT",
@@ -18,13 +19,18 @@
     }
   },
   "devDependencies": {
+    "@types/pino": "^6.3.11",
+    "@types/tap": "^15.0.5",
+    "@types/through2": "^2.0.36",
     "npm-watch": "^0.11.0",
     "pino": "^6.11.3",
     "pino-debug": "^2.0.0",
     "standard": "^8.6.0",
     "tap": "^12.0.1",
     "through2": "^4.0.2",
-    "typescript": "^4.2.4"
+    "tsd": "^0.17.0",
+    "ts-node": "^10.2.1",
+    "typescript": "^4.4.3"
   },
   "peerDependencies": {
     "pino": "*"
@@ -35,9 +41,11 @@
   "scripts": {
     "lint-fix": "standard --fix index.js examples/*js tests/*.js",
     "pretest": "npm run lint-fix",
-    "test": "tap tests/*.js",
+    "test": "tap tests/*.js && npm run test-ts && npm run test-types",
+    "test-ts": "ts-node tests/test-ts.ts",
     "test-watch": "tap tests/*.js tests/*.ts",
-    "test:ci": "npm run pretest && tap --node-arg=\"-r\" --node-arg=\"source-map-support/register\" tests/*.js tests/*.ts --coverage-report=lcovonly",
+    "test:ci": "npm run pretest && tap --node-arg=\"-r\" --node-arg=\"source-map-support/register\" tests/*.js tests/*.ts --coverage-report=lcovonly && npm run test-types",
+    "test-types": "tsc && tsd",
     "example": "env NODE_ENV=development node examples/index.js",
     "watch": "npm-watch"
   },
